+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-[Full diff](https://github.com/sider/runners/compare/0.46.1...HEAD)
+[Full diff](https://github.com/sider/runners/compare/0.47.0...HEAD)
+
+## 0.47.0
+
+[Full diff](https://github.com/sider/runners/compare/0.46.1...0.47.0)
 
 - Support `linter.<id>.dependencies` for Python tools [#2280](https://github.com/sider/runners/pull/2280)
 - **Flake8** Fix `Errno::ENOENT` on multiple configurations [#2281](https://github.com/sider/runners/pull/2281)